allow multiple -u USER options, accept any to unlock
The set of users allowed to unlock a the screen isn't necessarily the
same as a) the set of users with active open sessions, b) the owner of
the current VT, or c) the entire set of possible users on the machine.
A hard-coded list might work better, in practice, for me.  So the
password prompt will now accept any password from a list of
users.  root just becomes another user that can be added to the
command line, so those special cases go away.