get list of unlocking user passwords from logind/utmp
Now that we can unlock using a list of passwords, it makes sense to
allow any locally logged-in user's password to do so.  The list of
local users, obtained from logind (or else utmp, chosen at compile
time), are combined with user names specified on the commandline to
form a single list.